requirements.txt
cs50
Flask
pillow
pilgram